refactor: object into separate file
